+++ src/en/mangaowlto/build.gradle
@@ -1,6 +1,6 @@
 ext {
-    extName = 'MangaOwl'
-    extClass = '.MangaOwlFactory'
+    extName = 'MangaOwl.To'
+    extClass = '.MangaOwlToFactory'
     extVersionCode = 1
 }
