+++ src/en/mangaowlto/res/mipmap-mdpi/ic_launcher.png
