+++ src/en/mangaowlto/src/eu/kanade/tachiyomi/extension/en/mangaowlto/MangaOwlToFactory.kt
@@ -1,14 +1,14 @@
-package eu.kanade.tachiyomi.extension.en.mangaowl
+package eu.kanade.tachiyomi.extension.en.mangaowlto
 
 import eu.kanade.tachiyomi.source.Source
 import eu.kanade.tachiyomi.source.SourceFactory
 import eu.kanade.tachiyomi.source.model.Filter
 
-class MangaOwlFactory : SourceFactory {
+class MangaOwlToFactory : SourceFactory {
     class Genre(name: String, val uriPart: String) : Filter.CheckBox(name)
 
     override fun createSources(): List<Source> = listOf(
-        MangaOwl(
+        MangaOwlTo(
             collectionUrl = "comics",
             extraName = "Manga",
             genresList = listOf(
@@ -54,7 +54,7 @@ class MangaOwlFactory : SourceFactory {
                 Genre("Yuri", "6"),
             ),
         ),
-        MangaOwl(
+        MangaOwlTo(
             collectionUrl = "comics-new",
             extraName = "Comic",
             genresList = listOf(
